Extract Suit, Rank from Card
