+++ CardGame/CardGame/Rank.swift
@@ -1,0 +1,30 @@
+//
+//  Rank.swift
+//  CardGame
+//
+//  Created by 공명진 on 2019/09/29.
+//  Copyright © 2019 JK. All rights reserved.
+//
+
+import Foundation
+
+enum Rank: Int {
+    case ace = 1
+    case two, three, four, five, six, seven, eight, nine, ten
+    case jack, queen, king
+    func description() -> String {
+        switch self {
+        case .ace:
+            return "A"
+        case .jack:
+            return "J"
+        case .queen:
+            return "Q"
+        case .king:
+            return "K"
+        default:
+            return String(self.rawValue)
+        }
+    }
+}
+
